New folder structure, update cloudbuild.
