+++ avengers/frontend/webapp/next.config.mjs
