+++ avengers/frontend/webapp/src/app/components/assistantLogo.js
