+++ avengers/frontend/webapp/src/app/components/featuredBar.js
