+++ avengers/frontend/webapp/src/app/components/heroBar.js
